Separate click to select hook
